Double -> Int

+++ src/com/company/NumberInInterval.java
@@ -6,7 +6,7 @@ public class NumberInInterval {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        double number = Double.parseDouble(scanner.nextLine());
+        int number = Integer.parseInt(scanner.nextLine());
 
         if ( number != 0 && number >= -100 && number <= 100) {
             System.out.println("Yes");
